Bug 37334: Fix conflict with bug 39473
Bug 39473 added ^ and $ to options of the dropdown lists from the
column filters.

They are removed and replaced with "exact" later in the code (using
build_query).

But in the case of "_status", we do no get this code as it's not a
regular column (it does not have "data" attribute), but an additional
filters.

I don't think this fix is ideal, but we are too close to the release to
rework datatables.js

Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>