Bug 39875: (QA follow-up): Correct historycheck ui_context
This ui_context is supposed to be opac, not staff.
For the purposes of these tests, this is not important, as the ui_context is
only needed for the template params and this is testing if the history check
stage is supposed to show or not.
Reckon this was a copy paste without consequences hence missed, but the code
is now semantically more correct.

prove t/db_dependent/Koha/ILL/Request/Workflow/HistoryCheck.t

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>
(cherry picked from commit dfd028871ae0e35c177d0598e4e398db5c5f3975)
Signed-off-by: Paul Derscheid <paul.derscheid@lmscloud.de>