Bug 40221: Replace layout tables for component part display - OPAC
This patch updates the display of component parts on the bibliographic
detail pages in both the staff interface and OPAC. Table markup isn't
necessary to display simple rows of content.

To test, apply the patch and locate a record in the catalog with
component parts. (See Bug 11175 for sample record).

Look at the detail page for the record, under the components tab, in
both the OPAc and staff interface. In both cases the information should
be clearly displayed.

Sponsored-by: Athens County Public Libraries
Signed-off-by: Noah Tremblay <noah.tremblay@inlibro.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>