Bug 40022: Condense display by hiding versions under ellipses
This patch improves the contributor section display in the about page
by implementing a condensed view for version lists. When a contributor
has more than 2 versions for a role, only the first 2 are shown with
an ellipsis (...) that can be hovered to reveal the remaining versions.

The implementation:
- Sorts versions in reverse chronological order (most recent first)
- Shows first 2 versions directly
- Hides additional versions under a hoverable ellipsis using Bootstrap tooltips
- Adds CSS styling for the ellipsis with dotted underline and help cursor

This makes the contributor lists more readable while preserving access
to complete version information.

Signed-off-by: Paul Derscheid <paul.derscheid@lmscloud.de>
Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>