Bug 38201: DT - Remove 'autoWidth: false'
The default is "true" and is what we use in datatables.js.

This patch also moves the CSS rules to BaseResource to avoid
duplication.

The style of DT was modified in the item list of a train, because it inherited from fieldset.rows ol table (staff-global).

(cherry picked from commit 713d2598a05e501a46d1d1b8557e7d6bf8863f0b)
Signed-off-by: Michaela Sieber <michaela.sieber@kit.edu>
Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>