Bug 20601: Untranslatable strings in circulation statistics
This patch makes some updates to the circulation statistics wizard in
order to make more information translatable.

The patch also adds some template plugins to the output of filter
options in order to be able to show descriptions instead of codes, and
corrects the markup for the 'Patron library' dropdown which wasn't
working.

To test apply the patch and go to Reports -> Statistics wizards ->
Circulation.

- Test setting some or all options in the "Filter column" and running
  the report.
  - Confirm that the "Patron library" dropdown works.
- At the top of the results there should be a list of the filter options
  you selected and their values.
- All the filter labels should be correct, and the filter values should
  should be in their full/description form rather than in code form
  (e.g. library, item type, collection code, sorting field, etc.)

Sponsored-by: Athens County Public Libraries
Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Aleisha Amohia <aleishaamohia@hotmail.com>
Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>