Bug 40333: Tidy commit
The commit hook was deliverately disabled for the previous commit and
this commit added distinctly to ensure clarity between the functional
code change and the massive tidy that the hook created.

In that last commit we remove the last remaining TT variable from the
<script> block in that template and as such we unlock the JS tidying on
that block.. nice, except it hides the functional change in the mass
commit of whitespace, quoting and newline changes.

Sponsored-by: Open Fifth <https://openfith.co.uk>
Signed-off-by: Hannah Dunne-Howrie <hdunne-howrie@westminster.gov.uk>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>
Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>