Bug 40337: Handle enum in Koha::Patron->anonymize
We need to reset to the default value. We might have a problem (later)
if it's NOT NULL and no default is set, which is not handled by this
patch. But the situation does not seem to exist yet anyway.

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>
(cherry picked from commit b4a561c133ef8752e3af4e878a05d0588a27d468)
Signed-off-by: Paul Derscheid <paul.derscheid@lmscloud.de>